Moved some enemy settings to settings, refactored hovering enemy
Moved the file paths to settings and made them absolute paths that can
be found on any system. Added call to parent initializer in hovering
enemy and just pushed variables to that function and deleted
redundancies. Moved other variables to settings if needed. Changed x and
end to self in hovering enemy init incase changed further in parent
initializer.