Made parent class for screens, renamed graphics, refactored main loop
Renamed graphics.py to screens.py since file just holds screens. Made a
parent class for all screens in the file so we can use polymorphism in
main game loop. Switched from if elif cases for handling screens in main
loop to use a dictionary to select a screen using enumerations and
polymorphism.