Add files via upload
Created a highscores file containing the framework for our highscores. Still need to integrate it into pygame, as of right now it can only display to console the high scores. Need to update main to call functions in the file and send the players name and score to the addNewPlayer Function so their score can be added.