Add files via upload
Updated all files pertaining to screens and buttons. Added functionality for a new button called HighScores. Currently clicking the button leads to a blank page with a button to go back to the title screen. Also added a temporary button pmg file for High Scores (this was drawn in paint so a permanent replacement will soon be uploaded)