Add files via upload
High scores screen now displays a first score from the database and 4 placeholders. Still need to add a button to index different pages (scores past the 5th score) and add a way to keep track of which page the user is currently on so we can display scores past the fifth one.